# api/api.yaml
openapi: 3.0.1
info:
  version: 1.0.0
  title: Hacker News
  description: "This is the Open API documentation for our Hacker News clone's REST API, deployed at <https://arcane-journey-15614.herokuapp.com/>. <br><br><br>brrrr COMPLETAR!<br><br>All operations are executable. Only one operation requires authentication: `deleteTweet`. In this case, you must **Authorize** your request by providing the api_key vaule you got when you created the tweet."
servers:
  #heroku. comentar aquest i posar el vostre a la altra linea per a testejar a AWS
  - url: 'https://arcane-journey-15614.herokuapp.com/api/'
  #- url: 'http://3.82.7.163:8080/api/'  #inserir aqui la IP del vostre Cloud9
  #http://3.86.16.176/api/
paths:
  /news/{id}:
    get:
      tags:
      - news
      summary: Returns a specific new
      description: Returns a specific new, identified by id
      operationId: getInfoNew
      parameters:
      - name: id
        in: path
        description: ID of the new to return
        required: true  #not sure. potser tindrem un get /news per llistes? 
        schema:
          type: integer
          format: int64 #nose
      - name: token
        in: header
        description: user's auth. token (API key)
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/New'
        404:
          description: 'Error: New not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error404new'
        401:
          description: 'Unauthorized: Invalid API key or no API key provided'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error401'
...
  /news/upvoted:
    get:
      tags:
      - news
      summary: Returns a list of the news upvoted by the user
      description: Returns a list of the news upvoted by the user identified by the auth. token
      operationId: newsUpvotedByUser
      parameters:
      - name: token
        in: header
        description: user's auth. token (API key)
        required: true
        schema:
          type: string
          
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema: 
                type: array
                items:
                  $ref: '#/components/schemas/New'
        401:
          description: 'Unauthorized: Invalid API key or no API key provided'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error401'
...
  /comments/threads:
    get:
      tags:
      - comments
      summary: Returns a list of user comments 
      description: Returns a list user comments either by the authed user or by the specified user_id
      operationId: threads
      parameters:
        - name: token
          in: header
          description: user's auth. token (API key)
          required: true
          schema:
            type: string
        - name: userid
          in: query
          description: id of the user whose threads are to be obtained
          required: false
          schema:
            type: integer
            format: int64
            
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema: 
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        401:
          description: 'Unauthorized: Invalid API key or no API key provided'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error401'
        404:
          description: error. No user with that ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error404user'
  /comments/upvoted:
    get:
      tags:
      - comments
      summary: Returns a list of the comments upvoted by the user
      description: Returns a list of the comments upvoted by the user identified by the auth. token
      operationId: commentsUpvotedByUser
      parameters:
      - name: token
        in: header
        description: user's auth. token (API key)
        required: true
        schema:
          type: string
          
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema: 
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        401:
          description: 'Unauthorized: Invalid API key or no API key provided'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error401'
...
components:
  schemas:    #per quan vols mostrar un schema
    New:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 3
        title:
          type: string
          example: 'Amazing new title'
        url:
          type: string
          example: 'https://web.whatsapp.com/'
        text:
          type: string  #suposo q al json sera string. era de tipus text
          example: 'this is a text. hey'
        isurl:
          type: integer
          format: int64
          example: 1
        points:
          type: integer
          format: int64
          example: 70
        created_at:
          type: string
          format: date-time
          example: '2021-05-05 15:47:32.238268000 +0000'
        updated_at:
          type: string
          format: date-time
          example: '2021-05-05 15:47:32.238268000 +0000'
        user_id:
          type: integer
          format: int64
          example: 4
    Comment:
      type: object
      properties:
        id: 
          type: integer
          example: 2
        text:
          type: string
          example: "M'ha agradat molt la teva aportació al fòrum de l'assignatura. Aquí tens una píndola."
        created_at:
          type: string
          format: date-time
          example: '2021-05-05 15:47:32.238268000 +0000'
        updated_at:
          type: string
          format: date-time
          example: '2021-05-05 15:47:32.238268000 +0000'
        user_id:
          type: integer
          format: int64
          example: 4
        comment_id:
          type: integer
          example: 13
        new_id:
          type: integer
          example: 2
    ErrorResult:
      type: object
      properties:
        status:
          type: integer
          enum: [400, 401, 403, 404, 409]
        error: 
          type: string
          enum: ["Bad Request", "Unauthorized", "Forbidden", "Not Found", "Conflict"]
        message: 
          type: string
  examples:    #per quan vols mostrar un exemple
...
    error401:
      value:
        status: 401
        error: "Unauthorized"
        message: "Invalid API key"
...
    error404new:
      value:
        status: 404
        error: "Not Found"
        message: "No New with that ID"
...
    error404user:
      value:
        status: 404
        error: "Not Found"
        message: "No user with that ID"
...